Dont't unlink when it's not necessary, and warn when we do

+++ src/openpsa/setup/installer.php
@@ -146,7 +146,15 @@ class installer
     {
         if (is_link($linkname))
         {
-            unlink($linkname);
+            if (realpath($linkname) === $target)
+            {
+                return;
+            }
+            else
+            {
+                $io->write('Replacing <comment>' . realpath($linkname) . '</comment> with <info>' . $target . '</info>');
+                unlink($linkname);
+            }
         }
         if (!@symlink($target, $linkname))
         {
